fix model crud route

+++ Poseiden-skeleton/src/test/java/com/nnk/springboot/BidTests.java
@@ -38,7 +38,7 @@ public class BidTests {
 		Assert.assertTrue(listResult.size() > 0);
 
 		// Delete
-		Integer id = bid.getBidListId();
+		Integer id = Math.toIntExact(bid.getBidListId());
 		bidListRepository.delete(bid);
 		Optional<BidList> bidList = bidListRepository.findById(id);
 		Assert.assertFalse(bidList.isPresent());
