+++ Poseiden-skeleton/src/test/java/com/nnk/springboot/TradeTests.java
@@ -38,7 +38,7 @@ public class TradeTests {
 		Assert.assertTrue(listResult.size() > 0);
 
 		// Delete
-		Integer id = trade.getTradeId();
+		Integer id = Math.toIntExact(trade.getTradeId());
 		tradeRepository.delete(trade);
 		Optional<Trade> tradeList = tradeRepository.findById(id);
 		Assert.assertFalse(tradeList.isPresent());
